[APG-1041] Refactor security configuration and add custom JWT converter (#5)
